Add option to re-use single <style> element
This adds the ability for the style-loader to re-use a single <style>
element reference when adding/removing styles from the page. This is
particularly important when you want to use style-loader with a browser
like IE9, which enforces a hard-limit on the # of style tags allowed on
the page at any given point.

This is added as a new option `singleton`, and is automatically turned
on if the browser is determined to be IE9.