Better solution when combining HMR with singleton tags
Now, instead of assuming a constant length (d'oh), the boundaries of each
style block will be demarkated with a start/end token. When a style is updated,
the CSS block contents in between the boundaries is replaced. If a style is
removed, the entire section is removed. New sections are added to the bottom.