[#39] Don't use styleSheet to store state of css text
IE9 will actually drop certain things (like media queries) from the
cssText property when it is read back out. To fix, store the raw css
text in a separately managed array, and use that to form the entire
string that should be set on the style element's cssText property.

Fixes #39.