feat: add `prepend` query option
If defined, style-loader will prepend the css file instead of
appending it. Which means that the created `<style>` tag will be on
top of the `<head>` instead of at the bottom.

This allows you to build components with default styling that can be
overriden by user stylesheets.You can enable or disable it with the
`prepend` query parameter (`?prepend`).