docs(readme): Fix the `<style>` code block
PR #227 made the README not broken completely on the npm package page (https://www.npmjs.com/package/style-loader) but didn't make the code block a real code block. Backticks are only used to delimit code blocks when used on the top level of the Markdown file; once they appear in an HTML block they're treated as regular symbols. The proper fix is to use the `<code>` tag.