fix: Allow css-modules HMR when classes are unchanged
Fixes webpack-contrib/css-loader#669.

Previously, HMR was disabled if css-modules was enabled (in css-loader)
and at least one class was present in the file. This was because JS
files depend on the classes, and as such have to be reloaded when the
CSS changes in case the classes have changed.

Now, HMR _is_ used if no classes change. Othewise the whole page is
reloaded like before.